ci: prune per-commit CUDA tests — move 25 files + 13 testcases to test/manual/
Per-commit CUDA pipeline cut by ~174 minutes of registered est_time
(8622s whole-file moves + 1831s subset removals = 10453s total). 39
tests moved (25 whole files + 13 testcase classes) — all preserved as
runnable code under test/manual/. Registry parser only walks
test/registered/, so moved files are no longer auto-discovered by CI.

Subset removals scale est_time proportionally to remaining test classes
rounded up to nearest 10s. Removed classes archived to
test/manual/<sub>/test_<file>_archived.py with full module scope.

Source: 2026-05-07 working session
(~/Desktop/ci-pruning-2026-05-07_by-suite_v2.md, READ-ONLY).
