+++ test/manual/4-gpu-models/test_qwen3_next_models.py
