+++ test/manual/8-gpu-models/test_deepseek_v3_basic.py
