+++ test/manual/8-gpu-models/test_dsa_models_basic.py
