+++ test/manual/attention/test_local_attn.py
