+++ test/manual/core/test_gpt_oss_1gpu.py
