+++ test/manual/mla/test_flashmla_archived.py
@@ -1,53 +1,51 @@
+"""Archived test classes split out of test/registered/mla/test_flashmla.py.
+
+Originally registered with `register_cuda_ci(...)`. Moved here as part of
+the per-commit pruning effort to keep the code reachable manually.
+Run with `python3 test/manual/mla/test_flashmla_archived.py`.
+"""
+
+"""
+Usage:
+python3 test/registered/mla/test_flashmla.py
+"""
+
 import unittest
 from types import SimpleNamespace
 
-import requests
 import torch
 
 from sglang.srt.utils import kill_process_tree
-from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.run_eval import run_eval
 from sglang.test.test_utils import (
+    DEFAULT_MODEL_NAME_FOR_TEST_MLA,
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
     DEFAULT_URL_FOR_TEST,
-    CustomTestCase,
     popen_launch_server,
 )
 
-# FlashInfer MLA backend tests with MTP speculative decoding
-register_cuda_ci(est_time=130, suite="stage-b-test-1-gpu-large")
 
-
-class TestFlashinferMLAMTP(CustomTestCase):
+# FlashMLA attention backend tests with MTP speculative decoding
+class TestFlashMLAAttnBackend(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls.model = "lmsys/sglang-ci-dsv3-test"
+        cls.model = DEFAULT_MODEL_NAME_FOR_TEST_MLA
         cls.base_url = DEFAULT_URL_FOR_TEST
         other_args = ["--trust-remote-code"]
         if torch.cuda.is_available() and torch.version.cuda:
             other_args.extend(
                 [
                     "--cuda-graph-max-bs",
-                    "4",
-                    "--enable-torch-compile",
-                    "--torch-compile-max-bs",
-                    "1",
-                    "--speculative-algorithm",
-                    "EAGLE",
-                    "--speculative-num-steps",
-                    "3",
-                    "--speculative-eagle-topk",
-                    "1",
-                    "--speculative-num-draft-tokens",
-                    "4",
+                    "2",
                     "--attention-backend",
-                    "flashinfer",
+                    "flashmla",
                 ]
             )
+        # Use longer timeout for DeepGEMM JIT compilation which can take 10-20 minutes
         cls.process = popen_launch_server(
             cls.model,
             cls.base_url,
-            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
+            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH * 2,
             other_args=other_args,
         )
 
@@ -56,8 +54,6 @@ class TestFlashinferMLAMTP(CustomTestCase):
         kill_process_tree(cls.process.pid)
 
     def test_gsm8k(self):
-        requests.get(self.base_url + "/flush_cache")
-
         args = SimpleNamespace(
             base_url=self.base_url,
             model=self.model,
@@ -71,13 +67,6 @@ class TestFlashinferMLAMTP(CustomTestCase):
         print(metrics)
 
         self.assertGreater(metrics["score"], 0.60)
-
-        server_info = requests.get(self.base_url + "/server_info").json()
-        avg_spec_accept_length = server_info["internal_states"][0][
-            "avg_spec_accept_length"
-        ]
-        print(f"{avg_spec_accept_length=}")
-        self.assertGreater(avg_spec_accept_length, 2.5)
 
 
 if __name__ == "__main__":
