+++ test/manual/mla/test_mla_int8_deepseek_v3_archived.py
@@ -1,11 +1,16 @@
+"""Archived test classes split out of test/registered/mla/test_mla_int8_deepseek_v3.py.
+
+Originally registered with `register_cuda_ci(...)`. Moved here as part of
+the per-commit pruning effort to keep the code reachable manually.
+Run with `python3 test/manual/mla/test_mla_int8_deepseek_v3_archived.py`.
+"""
+
 import unittest
 from types import SimpleNamespace
 
-import requests
 import torch
 
 from sglang.srt.utils import kill_process_tree
-from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.run_eval import run_eval
 from sglang.test.test_utils import (
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
@@ -15,11 +20,9 @@ from sglang.test.test_utils import (
     popen_launch_server,
 )
 
+
 # DeepSeek-V3 INT8 quantization tests (channel and block INT8)
-register_cuda_ci(est_time=160, suite="stage-b-test-1-gpu-large")
-
-
-class TestDeepseekV3MTPChannelInt8(CustomTestCase):
+class TestMLADeepseekV3ChannelInt8(CustomTestCase):
     @classmethod
     def setUpClass(cls):
         cls.model = "lmsys/sglang-ci-dsv3-channel-int8-test"
@@ -33,16 +36,6 @@ class TestDeepseekV3MTPChannelInt8(CustomTestCase):
                     "--enable-torch-compile",
                     "--torch-compile-max-bs",
                     "2",
-                    "--speculative-algorithm",
-                    "EAGLE",
-                    "--speculative-draft-model-path",
-                    "lmsys/sglang-ci-dsv3-channel-int8-test-NextN",
-                    "--speculative-num-steps",
-                    "2",
-                    "--speculative-eagle-topk",
-                    "4",
-                    "--speculative-num-draft-tokens",
-                    "4",
                 ]
             )
         cls.process = popen_launch_server(
@@ -57,8 +50,6 @@ class TestDeepseekV3MTPChannelInt8(CustomTestCase):
         kill_process_tree(cls.process.pid)
 
     def test_gsm8k(self):
-        requests.get(self.base_url + "/flush_cache")
-
         args = SimpleNamespace(
             base_url=self.base_url,
             model=self.model,
@@ -71,18 +62,11 @@ class TestDeepseekV3MTPChannelInt8(CustomTestCase):
         metrics = run_eval(args)
         print(metrics)
 
-        self.assertGreater(metrics["score"], 0.60)
-
-        server_info = requests.get(self.base_url + "/server_info")
-        avg_spec_accept_length = server_info.json()["internal_states"][0][
-            "avg_spec_accept_length"
-        ]
-        print(f"{avg_spec_accept_length=}")
-        self.assertGreater(avg_spec_accept_length, 2.5)
+        self.assertGreaterEqual(metrics["score"], 0.61)
 
 
 @unittest.skipIf(is_in_ci(), "To reduce the CI execution time.")
-class TestDeepseekV3MTPBlockInt8(CustomTestCase):
+class TestMLADeepseekV3BlockInt8(CustomTestCase):
     @classmethod
     def setUpClass(cls):
         cls.model = "lmsys/sglang-ci-dsv3-block-int8-test"
@@ -96,14 +80,6 @@ class TestDeepseekV3MTPBlockInt8(CustomTestCase):
                     "--enable-torch-compile",
                     "--torch-compile-max-bs",
                     "2",
-                    "--speculative-algorithm",
-                    "EAGLE",
-                    "--speculative-num-steps",
-                    "2",
-                    "--speculative-eagle-topk",
-                    "4",
-                    "--speculative-num-draft-tokens",
-                    "4",
                 ]
             )
         cls.process = popen_launch_server(
@@ -118,8 +94,6 @@ class TestDeepseekV3MTPBlockInt8(CustomTestCase):
         kill_process_tree(cls.process.pid)
 
     def test_gsm8k(self):
-        requests.get(self.base_url + "/flush_cache")
-
         args = SimpleNamespace(
             base_url=self.base_url,
             model=self.model,
@@ -132,14 +106,7 @@ class TestDeepseekV3MTPBlockInt8(CustomTestCase):
         metrics = run_eval(args)
         print(metrics)
 
-        self.assertGreater(metrics["score"], 0.60)
-
-        server_info = requests.get(self.base_url + "/server_info")
-        avg_spec_accept_length = server_info.json()["internal_states"][0][
-            "avg_spec_accept_length"
-        ]
-        print(f"{avg_spec_accept_length=}")
-        self.assertGreater(avg_spec_accept_length, 2.5)
+        self.assertGreater(metrics["score"], 0.62)
 
 
 if __name__ == "__main__":
