+++ test/manual/models/test_nvidia_nemotron_nano_v2.py
