+++ test/manual/models/test_nvidia_nemotron_nano_v2_vl.py
