+++ test/manual/models/test_qwen_models.py
