+++ test/manual/perf/test_bench_one_batch_1gpu.py
