+++ test/manual/prefill_only/test_encoder_embedding_models.py
