+++ test/manual/quant/test_deepseek_v32_fp4_4gpu.py
