+++ test/manual/spec/eagle/test_eagle3_basic.py
