+++ test/registered/4-gpu-models/test_qwen3_next_models_mtp.py
@@ -6,35 +6,9 @@ from sglang.test.kits.kl_divergence_kit import KLDivergenceMixin
 from sglang.test.kits.prefix_cache_branching_kit import PrefixCacheBranchingMixin
 from sglang.test.server_fixtures.default_fixture import DefaultServerBase
 
-register_cuda_ci(est_time=422, suite="stage-c-test-4-gpu-h100")
+register_cuda_ci(est_time=290, suite="stage-c-test-4-gpu-h100")
 
 QWEN3_NEXT_MODEL = "Qwen/Qwen3-Next-80B-A3B-Instruct"
-
-
-class TestQwen3NextMTP(GSM8KMixin, KLDivergenceMixin, DefaultServerBase):
-    model = QWEN3_NEXT_MODEL
-    gsm8k_accuracy_thres = 0.93
-    kl_div_thres = 0.0025
-    other_args = [
-        "--trust-remote-code",
-        "--speculative-algorithm",
-        "NEXTN",
-        "--speculative-num-steps",
-        "3",
-        "--speculative-eagle-topk",
-        "1",
-        "--speculative-num-draft-tokens",
-        "4",
-        "--mem-fraction-static",
-        "0.8",
-        "--tp",
-        "4",
-        "--chunked-prefill-size",
-        "2048",
-        "--mamba-scheduler-strategy",
-        "no_buffer",
-        "--disable-radix-cache",
-    ]
 
 
 class TestQwen3NextMTPTopk(
