+++ test/registered/models/test_nvidia_nemotron_3_nano.py
@@ -5,7 +5,7 @@ from sglang.test.kits.lm_eval_kit import LMEvalMixin
 from sglang.test.server_fixtures.default_fixture import DefaultServerBase
 
 register_cuda_ci(
-    est_time=564,
+    est_time=190,
     suite="stage-b-test-2-gpu-large",
 )
 
@@ -16,30 +16,6 @@ NEMOTRON_3_NANO_THINKING_ARGS = [
     "--reasoning-parser",
     "deepseek-r1",
 ]
-
-
-class TestNvidiaNemotron3Nano30BBF16(LMEvalMixin, DefaultServerBase):
-    """Test Nemotron-3-Nano-30B BF16 model with lm-eval GSM8K evaluation."""
-
-    model = "nvidia/NVIDIA-Nemotron-3-Nano-30B-A3B-BF16"
-    model_config_name = "lm_eval_configs/NVIDIA-Nemotron-3-Nano-30B-A3B-BF16.yaml"
-    other_args = [
-        "--tp-size",
-        "2",
-    ] + NEMOTRON_3_NANO_THINKING_ARGS
-
-
-class TestNvidiaNemotron3Nano30BBF16FlashInfer(LMEvalMixin, DefaultServerBase):
-    """Test Nemotron-3-Nano-30B BF16 model with lm-eval GSM8K evaluation using flashinfer mamba backend."""
-
-    model = "nvidia/NVIDIA-Nemotron-3-Nano-30B-A3B-BF16"
-    model_config_name = "lm_eval_configs/NVIDIA-Nemotron-3-Nano-30B-A3B-BF16.yaml"
-    other_args = [
-        "--tp-size",
-        "2",
-        "--mamba-backend",
-        "flashinfer",
-    ] + NEMOTRON_3_NANO_THINKING_ARGS
 
 
 class TestNvidiaNemotron3Nano30BFP8(LMEvalMixin, DefaultServerBase):
