+++ test/registered/quant/test_awq.py
@@ -13,7 +13,7 @@ from sglang.test.test_utils import (
     popen_launch_server,
 )
 
-register_cuda_ci(est_time=226, suite="stage-b-test-1-gpu-large")
+register_cuda_ci(est_time=160, suite="stage-b-test-1-gpu-large")
 register_amd_ci(est_time=200, suite="stage-b-test-1-gpu-large-amd")
 
 
@@ -78,40 +78,6 @@ class TestAWQMarlinBfloat16(CustomTestCase):
 
         metrics = run_eval(args)
         self.assertGreater(metrics["score"], 0.83)
-
-
-@unittest.skipIf(is_in_amd_ci(), "AWQ Marlin is not supported on AMD GPUs")
-class TestAWQMarlinFloat16(CustomTestCase):
-    """
-    Verify that the model can be loaded with float16 dtype and awq_marlin quantization
-    """
-
-    @classmethod
-    def setUpClass(cls):
-        cls.model = "QuantTrio/Qwen3-VL-30B-A3B-Instruct-AWQ"
-        cls.base_url = DEFAULT_URL_FOR_TEST
-        cls.process = popen_launch_server(
-            cls.model,
-            cls.base_url,
-            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-            other_args=["--dtype", "float16", "--quantization", "awq_marlin"],
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        kill_process_tree(cls.process.pid)
-
-    def test_mmlu(self):
-        args = SimpleNamespace(
-            base_url=self.base_url,
-            model=self.model,
-            eval_name="mmlu",
-            num_examples=64,
-            num_threads=32,
-        )
-
-        metrics = run_eval(args)
-        self.assertGreater(metrics["score"], 0.85)
 
 
 if __name__ == "__main__":
